feat: add parent cache verification and a setting to enable it
When the cache files are generated, they are hashed and that hash is
persisted.  When the cache file is opened, it is optionally verified
for integrity via the 'verify_cache' setting (disabled by default).

When this code is run, if the checksum files do not exist, we
re-generate the parent cache file in order to write out that checksum
file since what's already on disk may be corrupted.